refactor(mgmt): Split table_extend.rs code
The code for building the stateless NAT tables has grown and contains
several logical sections. It would be more readable to have each of them
in a dedicated submodule, along with their relevant unit tests, so this
commit does just that.

The three portions of code are:

- The code to collapse the list of allowed and excluded prefixes into an
  equivalent list of allowed fragments only, moved to a new file
  collapse.rs
- The code to build the TrieValue entries for the NAT tables, where we
  split the prefix lists again into pairs of ranges with a 1:1 mapping,
  moved to a new file range_builder.rs
- The higher-level API function, add_peering(), and its direct
  dependency helper, that we move to the mod.rs

No functional change.
