+++ mgmt/src/models/internal/natconfig/collapse.rs
@@ -1,349 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Open Network Fabric Authors
 
-//! NAT rule tables entries creation
-
-use crate::models::external::overlay::vpc::{Peering, VpcTable};
-use crate::models::external::overlay::vpcpeering::{VpcExpose, VpcManifest};
-use nat::stateless::config::prefixtrie::{PrefixTrie, TrieError};
-use nat::stateless::config::tables::{NatPrefixRuleTable, PerVniTable, TrieValue};
-use net::vxlan::Vni;
-use routing::prefix::{Prefix, PrefixSize};
-use std::cmp::Ordering;
+use super::NatPeeringError;
+use crate::models::external::overlay::vpc::Peering;
+use routing::prefix::Prefix;
 use std::collections::BTreeSet;
-use std::fmt::Debug;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 use tracing::error;
-
-/// Error type for NAT peering table extension operations.
-#[derive(thiserror::Error, Debug)]
-pub enum NatPeeringError {
-    #[error("entry already exists")]
-    EntryExists,
-    #[error("failed to split prefix {0}")]
-    SplitPrefixError(Prefix),
-    #[error("malformed peering")]
-    MalformedPeering,
-}
-
-fn add_prefix_size(
-    offset: u128,
-    prefix_size: PrefixSize,
-    is_ipv4: bool,
-) -> Result<u128, NatPeeringError> {
-    match (is_ipv4, prefix_size) {
-        (true, PrefixSize::U128(size)) => {
-            if offset > u128::from(u32::MAX) - size {
-                // Adding the size of the current prefix to the offset would overflow the IP address
-                // space, which makes no sense. We have a malformed peering.
-                return Err(NatPeeringError::MalformedPeering);
-            }
-            Ok(offset + size)
-        }
-        (false, PrefixSize::U128(size)) => {
-            if offset > u128::MAX - size {
-                return Err(NatPeeringError::MalformedPeering);
-            }
-            Ok(offset + size)
-        }
-        // We've covered all existing addresses in the IPv6, but still haven't found our prefix.
-        // We have a malformed peering.
-        _ => Err(NatPeeringError::MalformedPeering),
-    }
-}
-
-fn add_offset_to_address(addr: &IpAddr, offset: PrefixSize) -> Result<IpAddr, NatPeeringError> {
-    match addr {
-        IpAddr::V4(addr) => {
-            let addr = u32::from(*addr)
-                + u32::try_from(
-                    u128::try_from(offset).map_err(|_| NatPeeringError::MalformedPeering)?,
-                )
-                .map_err(|_| NatPeeringError::MalformedPeering)?;
-            Ok(IpAddr::V4(addr.into()))
-        }
-        IpAddr::V6(addr) => {
-            let addr = u128::from(*addr)
-                + u128::try_from(offset).map_err(|_| NatPeeringError::MalformedPeering)?;
-            Ok(IpAddr::V6(addr.into()))
-        }
-    }
-}
-
-#[derive(Debug)]
-struct RangeBuilder<'a> {
-    vni: Vni,
-
-    prefix_iter_orig: std::collections::btree_set::Iter<'a, Prefix>,
-    prefix_iter_target: std::collections::btree_set::Iter<'a, Prefix>,
-
-    prefix_cursor_orig: Option<&'a Prefix>,
-    prefix_cursor_target: Option<&'a Prefix>,
-
-    addr_cursor_orig: Option<IpAddr>,
-    addr_cursor_target: Option<IpAddr>,
-
-    offset_cursor_orig: PrefixSize,
-    offset_cursor_target: PrefixSize,
-}
-
-impl<'a> RangeBuilder<'a> {
-    fn new(
-        vni: Vni,
-        prefixes_to_update: &'a BTreeSet<Prefix>,
-        prefixes_to_point_to: &'a BTreeSet<Prefix>,
-    ) -> Self {
-        let mut builder = Self {
-            vni,
-            prefix_iter_orig: prefixes_to_update.iter(),
-            prefix_iter_target: prefixes_to_point_to.iter(),
-            prefix_cursor_orig: None,
-            prefix_cursor_target: None,
-            addr_cursor_orig: None,
-            addr_cursor_target: None,
-            offset_cursor_orig: PrefixSize::U128(0),
-            offset_cursor_target: PrefixSize::U128(0),
-        };
-
-        builder.prefix_cursor_orig = builder.prefix_iter_orig.next();
-        builder.addr_cursor_orig = builder.prefix_cursor_orig.map(Prefix::as_address);
-
-        builder.prefix_cursor_target = builder.prefix_iter_target.next();
-        builder.addr_cursor_target = builder.prefix_cursor_target.map(Prefix::as_address);
-
-        builder
-    }
-}
-
-impl Iterator for RangeBuilder<'_> {
-    type Item = Result<TrieValue, NatPeeringError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.offset_cursor_orig >= PrefixSize::Ipv6MaxAddrs
-            || self.offset_cursor_target >= PrefixSize::Ipv6MaxAddrs
-        {
-            // We have covered the whole IPv6 address space, we have no reason to go any further.
-            return None;
-        }
-        // We've reached the end of the prefix list, the iterator is done.
-        let prefix_key = self.addr_cursor_orig?;
-
-        match (
-            self.prefix_cursor_orig,
-            self.prefix_cursor_target,
-            self.addr_cursor_orig,
-            self.addr_cursor_target,
-        ) {
-            (Some(orig_prefix), Some(target_prefix), Some(orig_addr), Some(target_addr)) => {
-                let orig_prefix_size = orig_prefix.size();
-                let target_prefix_size = target_prefix.size();
-
-                // Create new range based on current cursor values
-                let mut value = TrieValue {
-                    vni: Some(self.vni),
-                    orig_range_start: orig_addr,
-                    orig_range_end: orig_addr,
-                    target_range_start: target_addr,
-                };
-
-                // Determine next prefix
-                let prefix_orig_remain_size = orig_prefix_size - self.offset_cursor_orig;
-                let prefix_target_remain_size = target_prefix_size - self.offset_cursor_target;
-
-                match prefix_orig_remain_size.partial_cmp(&prefix_target_remain_size) {
-                    Some(Ordering::Less) => {
-                        let range_size = prefix_orig_remain_size;
-
-                        // Update return value's orig range end
-                        let Ok(new_range_end) = add_offset_to_address(&orig_addr, range_size - 1)
-                        else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        value.orig_range_end = new_range_end;
-
-                        // original range cursor update: advance to next orig prefix
-                        self.prefix_cursor_orig = self.prefix_iter_orig.next();
-                        self.addr_cursor_orig = self.prefix_cursor_orig.map(Prefix::as_address);
-                        self.offset_cursor_orig = PrefixSize::U128(0);
-
-                        // target range cursor update: advance to corresponding offset in current target prefix
-                        let Ok(new_addr) = add_offset_to_address(&target_addr, range_size) else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        let Ok(offset) = self.offset_cursor_target.try_into() else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        let Ok(new_cursor) =
-                            add_prefix_size(offset, range_size, target_prefix.is_ipv4())
-                        else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        self.addr_cursor_target = Some(new_addr);
-                        self.offset_cursor_target = new_cursor.into();
-                    }
-                    Some(Ordering::Greater) => {
-                        let range_size = prefix_target_remain_size;
-
-                        // Update return value's orig range end
-                        let Ok(new_range_end) = add_offset_to_address(&orig_addr, range_size - 1)
-                        else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        value.orig_range_end = new_range_end;
-
-                        // target range cursor update: advance to next target prefix
-                        self.prefix_cursor_target = self.prefix_iter_target.next();
-                        self.addr_cursor_target = self.prefix_cursor_target.map(Prefix::as_address);
-                        self.offset_cursor_target = PrefixSize::U128(0);
-
-                        // original range cursor update: advance to corresponding offset in current orig prefix
-                        let Ok(new_addr) = add_offset_to_address(&orig_addr, range_size) else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        let Ok(offset) = self.offset_cursor_orig.try_into() else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        let Ok(new_cursor) =
-                            add_prefix_size(offset, range_size, orig_prefix.is_ipv4())
-                        else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        self.addr_cursor_orig = Some(new_addr);
-                        self.offset_cursor_orig = new_cursor.into();
-                    }
-                    Some(Ordering::Equal) => {
-                        // Update return value's orig range end
-                        let Ok(new_range_end) =
-                            add_offset_to_address(&orig_addr, prefix_orig_remain_size - 1)
-                        else {
-                            return Some(Err(NatPeeringError::MalformedPeering));
-                        };
-                        value.orig_range_end = new_range_end;
-
-                        // original range cursor update: advance to next orig prefix
-                        self.prefix_cursor_orig = self.prefix_iter_orig.next();
-                        self.addr_cursor_orig = self.prefix_cursor_orig.map(Prefix::as_address);
-                        self.offset_cursor_orig = PrefixSize::U128(0);
-
-                        // target range cursor update: advance to next target prefix
-                        self.prefix_cursor_target = self.prefix_iter_target.next();
-                        self.addr_cursor_target = self.prefix_cursor_target.map(Prefix::as_address);
-                        self.offset_cursor_target = PrefixSize::U128(0);
-                    }
-                    None => {
-                        return Some(Err(NatPeeringError::MalformedPeering));
-                    }
-                }
-                Some(Ok(value))
-            }
-            // Both prefix lists have the same size and the cursor moves at the same speed, so we
-            // should reach the end of both lists at the same time. If we failed to retrieve the
-            // next prefix for one side only, this is a mistake.
-            (None, Some(_), _, _) | (Some(_), None, _, _) => {
-                Some(Err(NatPeeringError::MalformedPeering))
-            }
-            // We've cycled over both lists, we're done. (We should not reach this point, we should
-            // have returned at the top of the function.)
-            _ => None,
-        }
-    }
-}
-
-fn generate_trie_values<'a>(
-    vni: Vni,
-    prefixes_to_update: &'a BTreeSet<Prefix>,
-    prefixes_to_point_to: &'a BTreeSet<Prefix>,
-) -> impl Iterator<Item = Result<TrieValue, NatPeeringError>> {
-    RangeBuilder::<'a>::new(vni, prefixes_to_update, prefixes_to_point_to)
-}
-
-fn generate_public_trie_values(
-    vni: Vni,
-    expose: &VpcExpose,
-) -> impl Iterator<Item = Result<TrieValue, NatPeeringError>> {
-    generate_trie_values(vni, &expose.ips, &expose.as_range)
-}
-
-fn generate_private_trie_values(
-    vni: Vni,
-    expose: &VpcExpose,
-) -> impl Iterator<Item = Result<TrieValue, NatPeeringError>> {
-    generate_trie_values(vni, &expose.as_range, &expose.ips)
-}
-
-// Note: add_peering(table, peering) should be part of PerVniTable, but we prefer to keep it in a
-// separate submodule because it relies on definitions from the external models, unlike the rest of
-// the PerVniTable implementation.
-//
-/// Add a [`Peering`] to a [`PerVniTable`]
-///
-/// # Errors
-///
-/// Returns an error if some lists of prefixes contain duplicates
-pub fn add_peering(
-    table: &mut PerVniTable,
-    peering: &Peering,
-    vpc_table: &VpcTable,
-) -> Result<(), NatPeeringError> {
-    let new_peering = collapse_prefixes_peering(peering)?;
-
-    let mut local_expose_indices = vec![];
-
-    new_peering.local.exposes.iter().try_for_each(|expose| {
-        if expose.as_range.is_empty() {
-            // Nothing to do for source NAT, get out of here
-            return Ok(());
-        }
-        // Create new peering table for source NAT
-        let mut peering_table = NatPrefixRuleTable::new();
-
-        // For each private prefix, add an entry containing the set of public prefixes
-        generate_public_trie_values(table.vni, expose).try_for_each(|value| {
-            peering_table
-                .insert(&value?)
-                .map_err(|_| NatPeeringError::EntryExists)
-        })?;
-
-        // Add peering table to PerVniTable
-        table.src_nat_prefixes.push(peering_table);
-        local_expose_indices.push(table.src_nat_prefixes.len() - 1);
-        Ok(())
-    })?;
-
-    /* get vni for remote manifest */
-    let remote_vni = vpc_table
-        .get_vpc_by_vpcid(new_peering.remote_id)
-        .unwrap_or_else(|| unreachable!())
-        .vni;
-
-    // Update table for destination NAT
-    new_peering.remote.exposes.iter().try_for_each(|expose| {
-        // For each public prefix, add an entry containing the set of private prefixes
-        generate_private_trie_values(remote_vni, expose).try_for_each(|value| {
-            table
-                .dst_nat
-                .insert(&value?)
-                .map_err(|_| NatPeeringError::EntryExists)
-        })?;
-
-        // Update peering table to make relevant prefixes point to the new peering table, for each
-        // private prefix.
-        //
-        // Note that the public IPs are not always from the as_range list: if this list is empty,
-        // then there's no NAT required for the expose, meaning that the public IPs are those from
-        // the "ips" list.
-        let remote_public_prefixes = expose.public_ips();
-        remote_public_prefixes.iter().try_for_each(|prefix| {
-            table
-                .src_nat_peers
-                .rules
-                .insert(prefix, local_expose_indices.clone())
-                .map_err(|_| NatPeeringError::EntryExists)
-        })
-    })?;
-
-    Ok(())
-}
 
 // Collapse prefixes and exclusion prefixes in a Peering object: for each expose object, "apply"
 // exclusion prefixes to split allowed prefixes into smaller chunks, and remove exclusion prefixes
@@ -351,7 +14,7 @@ pub fn add_peering(
 //
 // For example, for a given expose with "ips" as 1.0.0.0/16 and "nots" as 1.0.0.0/18, the resulting
 // expose will contain 1.0.128.0/17 and 1.0.64.0/18 as "ips" prefixes, and an empty "nots" list.
-fn collapse_prefixes_peering(peering: &Peering) -> Result<Peering, NatPeeringError> {
+pub fn collapse_prefixes_peering(peering: &Peering) -> Result<Peering, NatPeeringError> {
     let mut clone = peering.clone();
     for expose in &mut clone
         .local
@@ -490,68 +153,10 @@ fn prefix_split(prefix: &Prefix) -> Result<(Prefix, Prefix), NatPeeringError> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::models::external::overlay::vpc::Vpc;
+    use crate::models::external::overlay::vpcpeering::{VpcExpose, VpcManifest};
     use ipnet::IpNet;
-    use iptrie::{IpRTrieSet, Ipv4Prefix, Ipv6Prefix};
-    use nat::stateless::config::tables::NatTables;
-    use net::vxlan::Vni;
+    use iptrie::IpRTrieSet;
     use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-    use std::str::FromStr;
-
-    #[test]
-    fn test_fabric() {
-        let expose1 = VpcExpose::empty()
-            .ip("1.1.0.0/16".into())
-            .not("1.1.5.0/24".into())
-            .not("1.1.3.0/24".into())
-            .not("1.1.1.0/24".into())
-            .ip("1.2.0.0/16".into())
-            .not("1.2.2.0/24".into())
-            .as_range("2.2.0.0/16".into())
-            .not_as("2.1.10.0/24".into())
-            .not_as("2.1.1.0/24".into())
-            .not_as("2.1.8.0/24".into())
-            .not_as("2.1.2.0/24".into())
-            .as_range("2.1.0.0/16".into());
-        let expose2 = VpcExpose::empty()
-            .ip("3.0.0.0/16".into())
-            .as_range("4.0.0.0/16".into());
-
-        let mut manifest1 = VpcManifest::new("VPC-1");
-        manifest1.add_expose(expose1).expect("Failed to add expose");
-        manifest1.add_expose(expose2).expect("Failed to add expose");
-
-        let expose3 = VpcExpose::empty()
-            .ip("1::/64".into())
-            .not("1::/128".into())
-            .as_range("1:1::/64".into())
-            .not_as("1:1::/128".into());
-        let expose4 = VpcExpose::empty()
-            .ip("2::/64".into())
-            .not("2::/128".into())
-            .as_range("2:4::/64".into())
-            .not_as("2:4::/128".into());
-
-        let mut manifest2 = VpcManifest::new("VPC-2");
-        manifest2.add_expose(expose3).expect("Failed to add expose");
-        manifest2.add_expose(expose4).expect("Failed to add expose");
-
-        let peering: Peering = Peering {
-            name: "test_peering".into(),
-            local: manifest1,
-            remote: manifest2,
-            remote_id: "12345".try_into().expect("Failed to create VPC ID"),
-        };
-
-        let vni = Vni::new_checked(100).unwrap();
-        let mut vpctable = VpcTable::new();
-        let mut vpc = Vpc::new("VPC", "12345", vni.as_u32()).unwrap();
-        vpc.peerings.push(peering.clone());
-        vpctable.add(vpc);
-
-        let mut vni_table = PerVniTable::new(vni);
-        add_peering(&mut vni_table, &peering, &vpctable).expect("Failed to build NAT tables");
-    }
 
     #[test]
     fn test_prefix_split() {
