+++ mgmt/src/processor/confbuild/internal.rs
@@ -19,7 +19,7 @@ use std::net::Ipv4Addr;
 use crate::models::external::gwconfig::{ExternalConfig, GwConfig};
 use crate::models::internal::InternalConfig;
 use crate::models::internal::interfaces::interface::InterfaceType;
-use crate::models::internal::natconfig::table_extend::add_peering;
+use crate::models::internal::natconfig::add_peering;
 use crate::models::internal::routing::bgp::{AfIpv4Ucast, AfL2vpnEvpn};
 use crate::models::internal::routing::bgp::{BgpConfig, BgpOptions, VrfImports};
 use crate::models::internal::routing::evpn::VtepConfig;
